Revert "Add/dashboard"
