+++ src/app/App.tsx
@@ -13,8 +13,7 @@ import FlashcardMenu from "../pages/FlashcardMenu";
 import CreateFlashcards from "../pages/CreateFlashcards";
 import MyFlashcards from "../pages/MyFlashcard/MyFlashcards";
 import FlashcardSet from "../pages/FlashcardSet/FlashcardSet";
-import PopularFlashcards from "../pages/PopularFlashcards/PopularFlashcards";
-import PopularFlashcardSet from "../pages/PopularFlashcards/PopularFlashcardSet";
+
 // router and routes
 const router = createBrowserRouter(
   createRoutesFromElements(
@@ -24,11 +23,6 @@ const router = createBrowserRouter(
       <Route path="/flashcards" element={<FlashcardMenu />} />
       <Route path="/create-flashcards" element={<CreateFlashcards />} />
       <Route path="/my-flashcards" element={<MyFlashcards />} />
-      <Route path="/popular-flashcards" element={<PopularFlashcards />} />
-      <Route
-        path="/popular-flashcards/:setId"
-        element={<PopularFlashcardSet />}
-      />
       <Route path="/flashcard-set" element={<FlashcardSet />} />
       <Route path="/my-flashcards/:setId" element={<FlashcardSet />} />
     </Route>
