+++ src/pages/MyFlashcard/MyFlashcards.tsx
@@ -1,6 +1,7 @@
 import { useEffect, useState } from "react";
 import { AddIcon } from "@chakra-ui/icons";
 import {
+  Button,
   Card,
   CardBody,
   CardFooter,
@@ -26,7 +27,6 @@ import { SimpleGrid } from "@chakra-ui/react";
 import { useNavigate } from "react-router-dom";
 import { SlOptionsVertical } from "react-icons/sl";
 import EditFlashcardSet from "./EditFlashcardSet";
-import { FaEye } from "react-icons/fa";
 
 interface FlashcardSet {
   id: string;
@@ -114,28 +114,32 @@ export default function MyFlashcards() {
             {flashcardSets.map((set) => (
               <Card
                 key={set.id}
-                bg="blue.50"
+                bg="white"
+                border="2px solid"
+                borderColor="blue.200"
                 borderRadius="lg"
                 boxShadow="lg"
                 transition="all 0.3s ease-in-out"
-                _hover={{ transform: "scale(1.05)", boxShadow: "2xl" }}
+                _hover={{
+                  transform: "scale(1.05)",
+                  boxShadow: "2xl",
+                  bg: "blue.100",
+                }}
               >
-                <CardHeader bg="blue.500" borderTopRadius="lg">
+                <CardHeader>
                   <HStack spacing={3}>
-                    <Heading fontSize="lg" color="white">
+                    <Heading fontSize="lg" color="blue.700">
                       {set.title}
                     </Heading>
                     <Spacer />
                     <HStack spacing={2}>
-                      <IconButton
-                        icon={<FaEye />}
-                        size="xs"
-                        color="blue.600"
-                        bg="transparent"
-                        aria-label="View"
+                      <Button
+                        size="sm"
+                        colorScheme="blue"
                         onClick={() => navigate(`/my-flashcards/${set.id}`)}
-                        _hover={{ color: "white", bg: "transparent" }}
-                      />
+                      >
+                        View
+                      </Button>
                       <Menu placement="bottom-end">
                         <MenuButton
                           as={IconButton}
@@ -143,7 +147,6 @@ export default function MyFlashcards() {
                           icon={<SlOptionsVertical />}
                           colorScheme="teal"
                           aria-label="Options"
-                          bg="transparent"
                         />
                         <MenuList>
                           <MenuItem onClick={() => openEditModal(set)}>
