--- src/pages/PopularFlashcards/PopularFlashcardsContent.tsx
@@ -1,241 +1,0 @@
-export const popularFlashcardContent = [
-  {
-    id: "1",
-    title: "Thermodynamics",
-    description:
-      "Explore the principles of energy transfer, heat, and work in physical systems.",
-    lastUpdated: "2024-01-25",
-    views: Math.floor(Math.random() * 10000),
-    flashcards: [
-      {
-        definition: "What is the definition of thermodynamics?",
-        answer:
-          "Thermodynamics is the study of energy and its transformations.",
-      },
-      {
-        definition: "What is the definition of thermodynamics?",
-        answer:
-          "Thermodynamics is the study of energy and its transformations.",
-      },
-    ],
-  },
-  {
-    id: "2",
-    title: "Material Science",
-    description:
-      "Understand the properties, structures, and performance of various materials.",
-    lastUpdated: "2024-01-03",
-    views: Math.floor(Math.random() * 10000),
-    flashcards: [
-      {
-        definition: "What is material science?",
-        answer:
-          "Material science is the study of the properties and applications of materials of construction or manufacture (such as ceramics, metals, polymers, and composites).",
-      },
-      {
-        definition: "What are polymers?",
-        answer:
-          "Polymers are large molecules composed of repeated subunits, known for their elasticity and plasticity.",
-      },
-    ],
-  },
-  {
-    id: "3",
-    title: "Quantum Mechanics",
-    description:
-      "Dive into the study of matter and energy at atomic and subatomic levels.",
-    lastUpdated: "2024-01-25",
-    views: Math.floor(Math.random() * 10000),
-    flashcards: [
-      {
-        definition: "What is quantum mechanics?",
-        answer:
-          "Quantum mechanics is a fundamental theory in physics that provides a description of the physical properties of nature at the scale of atoms and subatomic particles.",
-      },
-      {
-        definition: "What is the uncertainty principle?",
-        answer:
-          "The uncertainty principle is a fundamental concept in quantum mechanics stating that the position and momentum of a particle cannot both be precisely determined at the same time.",
-      },
-    ],
-  },
-  {
-    id: "4",
-    title: "Basic Math",
-    description:
-      "Learn essential mathematics concepts and formulas for everyday calculations.",
-    lastUpdated: "2024-01-04",
-    views: Math.floor(Math.random() * 10000),
-    flashcards: [
-      {
-        definition: "What is basic math?",
-        answer:
-          "Basic math includes arithmetic operations like addition, subtraction, multiplication, and division.",
-      },
-      {
-        definition: "What is a fraction?",
-        answer:
-          "A fraction represents a part of a whole or, more generally, any number of equal parts.",
-      },
-    ],
-  },
-  {
-    id: "5",
-    title: "World History",
-    description:
-      "Discover key historical events and dates that shaped the world.",
-    lastUpdated: "2024-01-16",
-    views: Math.floor(Math.random() * 10000),
-    flashcards: [
-      {
-        definition: "What is world history?",
-        answer:
-          "World history is the field of historical study that examines history from a global perspective.",
-      },
-      {
-        definition: "What was the Renaissance?",
-        answer:
-          "The Renaissance was a period in European history marking the transition from the Middle Ages to modernity, characterized by an interest in classical scholarship and values.",
-      },
-    ],
-  },
-  {
-    id: "6",
-    title: "English Grammar",
-    description:
-      "Master common grammar rules and examples to improve your writing skills.",
-    lastUpdated: "2024-01-17",
-    views: Math.floor(Math.random() * 10000),
-    flashcards: [
-      {
-        definition: "What is English grammar?",
-        answer:
-          "English grammar is the set of structural rules governing the composition of clauses, phrases, and words in the English language.",
-      },
-      {
-        definition: "What is a noun?",
-        answer:
-          "A noun is a word that functions as the name of a specific object or set of objects, such as living creatures, places, actions, qualities, states of existence, or ideas.",
-      },
-    ],
-  },
-  {
-    id: "7",
-    title: "Advanced Mathematics",
-    description: "Explore complex mathematical theories and applications.",
-    lastUpdated: "2024-01-17",
-    views: Math.floor(Math.random() * 10000),
-    flashcards: [
-      {
-        definition: "What is advanced mathematics?",
-        answer:
-          "Advanced mathematics includes complex topics such as calculus, linear algebra, and differential equations.",
-      },
-      {
-        definition: "What is calculus?",
-        answer:
-          "Calculus is a branch of mathematics that studies continuous change, through derivatives and integrals.",
-      },
-    ],
-  },
-  {
-    id: "8",
-    title: "Electronics",
-    description:
-      "Understand the fundamentals of electronic circuits and devices.",
-    lastUpdated: "2024-01-09",
-    views: Math.floor(Math.random() * 10000),
-    flashcards: [
-      {
-        definition: "What is electronics?",
-        answer:
-          "Electronics is the branch of physics and technology concerned with the design of circuits using transistors and microchips, and with the behavior and movement of electrons in a semiconductor, conductor, vacuum, or gas.",
-      },
-      {
-        definition: "What is a transistor?",
-        answer:
-          "A transistor is a semiconductor device used to amplify or switch electronic signals and electrical power.",
-      },
-    ],
-  },
-  {
-    id: "9",
-    title: "Microelectronics",
-    description:
-      "Learn about the design and function of microelectronic components.",
-    lastUpdated: "2024-01-06",
-    views: Math.floor(Math.random() * 10000),
-    flashcards: [
-      {
-        definition: "What is microelectronics?",
-        answer:
-          "Microelectronics is a subfield of electronics that deals with very small and microscopic electronic components.",
-      },
-      {
-        definition: "What is an integrated circuit?",
-        answer:
-          "An integrated circuit is a set of electronic circuits on one small flat piece (or 'chip') of semiconductor material, normally silicon.",
-      },
-    ],
-  },
-  {
-    id: "10",
-    title: "Software Development",
-    description:
-      "Gain insights into software development processes and best practices.",
-    lastUpdated: "2024-01-17",
-    views: Math.floor(Math.random() * 10000),
-    flashcards: [
-      {
-        definition: "What is software development?",
-        answer:
-          "Software development is the process of conceiving, specifying, designing, programming, documenting, testing, and bug fixing involved in creating and maintaining applications, frameworks, or other software components.",
-      },
-      {
-        definition: "What is agile methodology?",
-        answer:
-          "Agile methodology is a practice that promotes continuous iteration of development and testing throughout the software development lifecycle of the project.",
-      },
-    ],
-  },
-  {
-    id: "11",
-    title: "Differential Equations",
-    description:
-      "Study the methods for solving differential equations and their applications.",
-    lastUpdated: "2024-01-08",
-    views: Math.floor(Math.random() * 1000),
-    flashcards: [
-      {
-        definition: "What is a differential equation?",
-        answer:
-          "A differential equation is a mathematical equation that relates some function with its derivatives.",
-      },
-      {
-        definition: "What is a partial differential equation?",
-        answer:
-          "A partial differential equation is a differential equation that contains unknown multivariable functions and their partial derivatives.",
-      },
-    ],
-  },
-  {
-    id: "12",
-    title: "Trigonometry",
-    description:
-      "Understand the relationships between the angles and sides of triangles.",
-    lastUpdated: "2024-01-17",
-    views: Math.floor(Math.random() * 1000),
-    flashcards: [
-      {
-        definition: "What is trigonometry?",
-        answer:
-          "Trigonometry is a branch of mathematics that studies relationships involving lengths and angles of triangles.",
-      },
-      {
-        definition: "What is the sine function?",
-        answer:
-          "The sine function is a trigonometric function of an angle, defined as the ratio of the length of the opposite side to the hypotenuse in a right-angled triangle.",
-      },
-    ],
-  },
-];
