+++ src/shared/ui/Navbar.tsx
@@ -1,4 +1,4 @@
-import { CheckIcon } from "@chakra-ui/icons";
+import { CheckIcon, SearchIcon } from "@chakra-ui/icons";
 import {
   Avatar,
   Button,
@@ -10,9 +10,11 @@ import {
   useToast,
   AvatarBadge,
   Image,
+  InputGroup,
+  Input,
+  InputLeftElement,
 } from "@chakra-ui/react";
 import { useNavigate } from "react-router-dom";
-import SearchBar from "./SearchBar";
 
 export default function Navbar() {
   const toast = useToast();
@@ -27,11 +29,10 @@ export default function Navbar() {
     });
   };
   const navigate = useNavigate();
-
   return (
     <Flex as="nav" alignItems="center" mb="20px" gap="10px">
       <Image
-        src="src/shared/assets/Logo-Blue.png"
+        src="src/shared/assets/1.png"
         alt="BlueCards"
         boxSize="40px"
         onClick={() => navigate("/")}
@@ -43,7 +44,21 @@ export default function Navbar() {
 
       <Spacer />
 
-      <SearchBar />
+      <InputGroup maxW="400px" mx="auto">
+        <InputLeftElement pointerEvents="none">
+          <SearchIcon color="gray.400" />
+        </InputLeftElement>
+        <Input
+          type="text"
+          placeholder="Search flashcards..."
+          bg="white"
+          border="2px solid"
+          borderColor="blue.400"
+          focusBorderColor="blue.600"
+          borderRadius="2xl"
+          _placeholder={{ color: "gray.500" }}
+        />
+      </InputGroup>
 
       <Spacer />
